+++ tests/test_extensions/test_ext_autodoc_events.py
@@ -58,6 +58,26 @@ def test_cut_lines(app):
     ]
 
 
+def test_cut_lines_no_objtype():
+    docstring_lines = [
+        'first line',
+        '---',
+        'second line',
+        '---',
+        'third line ',
+        '',
+    ]
+    process = cut_lines(2)
+
+    process(None, 'function', 'func', None, {}, docstring_lines)  # type: ignore[arg-type]
+    assert docstring_lines == [
+        'second line',
+        '---',
+        'third line ',
+        ''
+    ]
+
+
 @pytest.mark.sphinx('html', testroot='ext-autodoc')
 def test_between(app):
     app.connect('autodoc-process-docstring', between('---', ['function']))