+++ sphinx/application.py
@@ -176,6 +176,7 @@ class Sphinx:
         self._fresh_env_used: bool | None = None
         self.extensions: dict[str, Extension] = {}
         self.registry = SphinxComponentRegistry()
+        self.project_class = Project  # can be modified by extensions in setup
 
         # validate provided directories
         self.srcdir = _StrPath(srcdir).resolve()
@@ -280,7 +281,7 @@ class Sphinx:
         self.events.emit('config-inited', self.config)
 
         # create the project
-        self.project = Project(self.srcdir, self.config.source_suffix)
+        self.project = self.project_class(self.srcdir, self.config.source_suffix)
 
         # set up the build environment
         self.env = self._init_env(freshenv)